Restructure market-closure tests
- Moved previous integration tests to e2e
- Moved root tests to unit
